fix(renovate-config): Set timezone to UTC and schedule to the middle of the working week
By default Renovate uses the timezone of the machine it's running on. This fixes it to UTC and changes the schedule to be the middle of the working week. Our repos can override this depending upon how far upstream they are in the repo dependency chain. See https://docs.renovatebot.com/faq/#control-renovates-schedule